Refactored investment interactions form handlers into own middleware
